% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{at_creat}
\alias{at_creat}
\title{Create an automatic tuning machine for mlr3.}
\usage{
at_creat(tuner, learner, search_space, inner_resample, measure)
}
\arguments{
\item{tuner}{\code{\link[mlr3tuning]{auto_tuner}}}

\item{learner}{\code{\link[mlr3tuning]{auto_tuner}}}

\item{search_space}{\code{\link[mlr3tuning]{auto_tuner}}}

\item{inner_resample}{\code{\link[mlr3tuning]{auto_tuner}}}

\item{measure}{\code{\link[mlr3tuning]{auto_tuner}}}
}
\value{
The AutoTuner wraps a mlr3::Learner and augments it with an automatic tuning process for a given set of hyperparameters.
}
\description{
Create an automatic tuning machine for mlr3.
}
\examples{
\donttest{
    library(mlr3verse)
    measure <- msr("classif.auc")
    inner_resample <- rsmp("cv", folds = 5)
    outer_resample <- rsmp("cv", folds = 5)
    tuner <- tnr("grid_search",resolution = 5)
    learner_rpart = lrn("classif.rpart",predict_type = "prob",id ="Classification_Tree")
    search_space_rpart = ps(cp = p_dbl(lower = 0.001, upper = 0.1))
    at_rpart = at_creat(tuner,learner_rpart,search_space_rpart,inner_resample,measure)
    }
}
