% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{Comp_imp}
\alias{Comp_imp}
\title{Integrate the variable (gene) importance evaluation across 10 different machine learning models.}
\usage{
Comp_imp(
  gene_exp,
  result_type = "all",
  num = 5,
  inner_resampless = 5,
  outer_resampless = 5,
  tuner_resolutions = 5,
  multi_threaded = T
)
}
\arguments{
\item{gene_exp}{For a gene expression matrix, it is required to have sample names as row names and gene names as column names.
Additionally, there should be an additional column called "type" consisting of 0 and 1, representing whether a sample has been treated or not.
Here, 0 represents the control group (untreated), and 1 represents the treated group.}

\item{result_type}{For a character variable, the parameter "result_type" can be one of the following:
"table," "plot," "gene," "gene_model," or "all."
If "result_type" is set to "table,"it will return a data frame that records all the data from training 10 different models.
If "result_type" is set to "plot," it will return a plotted image.
If "result_type" is set to "gene," it will return the filtered gene names.
If "result_type" is set to "gene_model," it will return the filtered gene names followed by the model.
If the parameter "all" is used, it will return a list containing the results for the previous four options.}

\item{num}{For a numeric variable，Specify the number of top-ranked variables to select for filtering.}

\item{inner_resampless}{For a character variable,\code{\link[mlr3]{rsmp}}}

\item{outer_resampless}{For a character variable,\code{\link[mlr3]{rsmp}}}

\item{tuner_resolutions}{For a character variable\code{\link[mlr3tuning]{tnr}}}

\item{multi_threaded}{A logic variable to enable multi-threading during execution.}
}
\value{
Return a list containing four elements: "table," "plot," "gene," and "gene_model." In the "table" element,
you will find the specific results of the 10 models across multiple resampling iterations.
The "plot" element will contain the gene importance plot. The "gene" element will include the filtered gene names.
And the "gene_model" element will provide the filtered gene names followed by the model name.
}
\description{
Integrate the variable (gene) importance evaluation across 10 different machine learning models.
}
\examples{
\donttest{
    data(test)
    res = Comp_imp(test[c(1:400),c(1:9)])
    res$plot
     }
}
