% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC.R
\name{COMP_ROC}
\alias{COMP_ROC}
\title{ROC curves across 10 different machine learning models.}
\usage{
COMP_ROC(
  gene_exp,
  multi_threaded = T,
  split_ratios = 0.7,
  inner_resampless = 5,
  outer_resampless = 5,
  tuner_resolutions = 5
)
}
\arguments{
\item{gene_exp}{For a gene expression matrix, it is required to have sample names as row names and gene names as column names.
Additionally, there should be an additional column called "type" consisting of 0 and 1, representing whether a sample has been treated or not.
Here, 0 represents the control group (untreated), and 1 represents the treated group.}

\item{multi_threaded}{A logic variable to enable multi-threading during execution.}

\item{split_ratios}{A numeric variable representing the ratio used to split the data into a testing set and a training set;
it indicates the proportion of the total data that will be allocated to the testing set.}

\item{inner_resampless}{\code{\link[mlr3]{rsmp}}}

\item{outer_resampless}{\code{\link[mlr3]{rsmp}}}

\item{tuner_resolutions}{\code{\link[mlr3tuning]{tnr}}}
}
\value{
ROC curve
}
\description{
ROC curves across 10 different machine learning models.
}
\examples{
\donttest{
    data(test)
    ROC = COMP_ROC(test[c(1:400),c(1:9)])
    ROC
    }
}
