% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{data_handle}
\alias{data_handle}
\title{Standardize Data}
\usage{
data_handle(data)
}
\arguments{
\item{data}{For a gene expression matrix, it is required to have sample names as row names and gene names as column names.
Additionally, there should be an additional column called "type" consisting of 0 and 1, representing whether a sample has been treated or not.
Here, 0 represents the control group (untreated), and 1 represents the treated group.}
}
\value{
Return the processed data.
}
\description{
Change "-" in column names to "_",
Convert the type to a factor variable.
}
\examples{
\donttest{
    data(test)
    data = data_handle(test)}
}
