% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency.R
\name{efficiency_mod}
\alias{efficiency_mod}
\title{Integrate the model performance evaluation across 10 different machine learning models.}
\usage{
efficiency_mod(
  gene_exp,
  result_type = "all",
  inner_resampless = 5,
  outer_resampless = 5,
  tuner_resolutions = 5,
  finial_score_msrs = c("classif.auc", "classif.ce", "classif.acc", "classif.precision",
    "classif.recall", "classif.sensitivity", "classif.specificity"),
  multi_threaded = T
)
}
\arguments{
\item{gene_exp}{For a gene expression matrix, it is required to have sample names as row names and gene names as column names.
Additionally, there should be an additional column called "type" consisting of 0 and 1, representing whether a sample has been treated or not.
Here, 0 represents the control group (untreated), and 1 represents the treated group.}

\item{result_type}{For a character variable, the parameter "result_type" can be one of the following:
"data," "plot," or "all."
If "result_type" is set to "data,"it will return a data frame that records all the data from training 10 different models.
If "result_type" is set to "plot," it will return a plotted image.
If the parameter "all" is used, it will return a list containing the results for the previous two options.}

\item{inner_resampless}{For a character variable,\code{\link[mlr3]{rsmp}}}

\item{outer_resampless}{For a character variable,\code{\link[mlr3]{rsmp}}}

\item{multi_threaded}{A logic variable to enable multi-threading during execution.}

\item{tuner_resolution}{\code{\link[mlr3tuning]{tnr}}}

\item{finial_score_msr}{A character vector representing performance metrics for the model evaluation,
including "classif.auc," "classif.ce," "classif.acc," "classif.precision,"
"classif.recall," "classif.sensitivity," and "classif.specificity."}
}
\value{
Return a list containing four elements: "data","plot" In the "data" element,
you will find the specific results of the 10 models across multiple resampling iterations.
The "plot" element will contain the model performance plot.
}
\description{
Integrate the model performance evaluation across 10 different machine learning models.
}
\examples{
\donttest{
    data(test)
    res = efficiency_mod(test[,c(1:10)])
    res$plot
    res$data[c(1:10),]
     }
}
