% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC.R
\name{ROC_lda}
\alias{ROC_lda}
\title{Generate the raw ROC data after training a machine learning model using LDA model
and making predictions.}
\usage{
ROC_lda(
  task,
  data,
  split_ratio = 0.7,
  inner_resamples = 5,
  outer_resamples = 5,
  tuner_resolution = 5
)
}
\arguments{
\item{task}{\code{\link[mlr3]{resample}}}

\item{data}{Result of data_handle().For a gene expression matrix, it is required to have sample names as row names and gene names as column names.
Additionally, there should be an additional column called "type" consisting of 0 and 1, representing whether a sample has been treated or not.
Here, 0 represents the control group (untreated), and 1 represents the treated group.}

\item{split_ratio}{A numeric variable representing the ratio used to split the data into a testing set and a training set;
it indicates the proportion of the total data that will be allocated to the testing set.}

\item{inner_resamples}{\code{\link[mlr3]{rsmp}}}

\item{outer_resamples}{\code{\link[mlr3]{rsmp}}}

\item{tuner_resolution}{\code{\link[mlr3tuning]{tnr}}}
}
\value{
If the data contained any NA value and na.rm=FALSE, NA is returned.
Otherwise, if smooth=FALSE, a list of class “roc” with the following fields:\code{\link[pROC]{roc}}
}
\description{
Generate the raw ROC data after training a machine learning model using LDA model
and making predictions.
}
\examples{
\donttest{
    data(test)
    data = data_handle(test[c(1:400),c(1:9)])
    task = task_creat(data)
    lda = ROC_lda(task,data)
    lda
    }
}
