% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROC.R
\name{roc_plot}
\alias{roc_plot}
\title{Plot the ROC curves using the ROC results.}
\usage{
roc_plot(roc)
}
\arguments{
\item{roc}{\code{\link[pROC]{roc}} The results of the functions, ROC_rpart, ROC_knn, ROC_svm, ROC_xgboost, ROC_rr,
ROC_glmnet, ROC_lda, ROC_logreg, ROC_naive_bayes, and ROC_nnet, with class "roc".}
}
\value{
ROC curve
}
\description{
Plot the ROC curves using the ROC results.
}
\examples{
\donttest{
    data(test)
    data = data_handle(test[c(1:400),c(1:9)])
    task = task_creat(data)
    rpart = ROC_rpart(task,data)
    roc_plot(rpart)
    }
}
