% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{svm_imp}
\alias{svm_imp}
\title{Obtain variable importance using a SVM model.}
\usage{
svm_imp(
  task,
  data,
  inner_resamples = 5,
  outer_resample = 5,
  tuner_resolution = 5,
  Measure = "classif.acc"
)

svm_imp(
  task,
  data,
  inner_resamples = 5,
  outer_resample = 5,
  tuner_resolution = 5,
  Measure = "classif.acc"
)
}
\arguments{
\item{task}{\code{\link[mlr3]{resample}}}

\item{data}{Result of data_handle().For a gene expression matrix, it is required to have sample names as row names and gene names as column names.
Additionally, there should be an additional column called "type" consisting of 0 and 1, representing whether a sample has been treated or not.
Here, 0 represents the control group (untreated), and 1 represents the treated group.}

\item{inner_resamples}{\code{\link[mlr3]{rsmp}}}

\item{outer_resample}{\code{\link[mlr3]{rsmp}}}

\item{tuner_resolution}{\code{\link[mlr3tuning]{tnr}}}

\item{Measure}{The default value is set to "classif.acc."\code{\link[mlr3]{Measure}}A character vector that can include any of the following metrics is used to evaluate model performance: "classif.acc,"
"classif.auc," "classif.bacc," "classif.bbrier," "classif.ce," "classif.costs," "classif.dor,"
"classif.precision," "classif.recall," "classif.sensitivity," "classif.specificity," "classif.tn,"
and "classif.tp." These metrics provide a comprehensive evaluation of the model's effectiveness.}
}
\value{
A data.frame records the dropout_loss of a model for each sampling operation,
which represents the degree of impact of removing a variable on the model.
The evaluation is done using RMSE (Root Mean Square Error)

A data.frame records the dropout_loss of a model for each sampling operation,
which represents the degree of impact of removing a variable on the model.
The evaluation is done using RMSE (Root Mean Square Error)
}
\description{
Obtain variable importance using a SVM model.

Obtain variable importance using a SVM model.
}
\examples{
\donttest{
    data(test)
    data = data_handle(test)
    task = task_creat(data)
    SVM = svm_imp(task,data,
        inner_resamples = 5,
        outer_resample = 5,
        tuner_resolution = 5)
    SVM[c(1:10),]
     }
\donttest{
    data(test)
    data = data_handle(test)
    task = task_creat(data)
    SVM = svm_imp(task,data,
        inner_resamples = 5,
        outer_resample = 5,
        tuner_resolution = 5,
        Measure = "classif.acc")
    SVM[c(1:10),]
     }
}
