% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency.R
\name{svm_model}
\alias{svm_model}
\title{Performance evaluation of SVM model}
\usage{
svm_model(
  task,
  data,
  inner_resamples = 5,
  outer_resamples = 5,
  tuner_resolution = 5,
  finial_score_msr = c("classif.auc", "classif.ce", "classif.acc", "classif.precision",
    "classif.recall", "classif.sensitivity", "classif.specificity")
)
}
\arguments{
\item{task}{A task of mlr3 \code{\link[mlr3]{resample}}}

\item{data}{Result of data_handle().For a gene expression matrix, it is required to have sample names as row names and gene names as column names.
Additionally, there should be an additional column called "type" consisting of 0 and 1, representing whether a sample has been treated or not.
Here, 0 represents the control group (untreated), and 1 represents the treated group.}

\item{inner_resamples}{\code{\link[mlr3]{rsmp}}}

\item{tuner_resolution}{\code{\link[mlr3tuning]{tnr}}}

\item{finial_score_msr}{A character vector representing performance metrics for the model evaluation,
including "classif.auc," "classif.ce," "classif.acc," "classif.precision,"
"classif.recall," "classif.sensitivity," and "classif.specificity."}

\item{outer_resample}{\code{\link[mlr3]{rsmp}}}
}
\value{
A data.frame containing the values of specific performance metrics
for each iteration of model training.
}
\description{
Performance evaluation of SVM model
}
\examples{
\donttest{
    data(test)
    data = data_handle(test)
    task = task_creat(data)
    result = svm_model(task,data,
        inner_resamples = 5,
        outer_resample = 5,
        tuner_resolution = 5,
        finial_score_msr = c("classif.auc","classif.ce","classif.acc",
        "classif.precision","classif.recall","classif.sensitivity",
         "classif.specificity")
        )
    result[c(1:5),]
     }
}
