% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{task_creat}
\alias{task_creat}
\title{Create Task}
\usage{
task_creat(data, target = "type")
}
\arguments{
\item{data}{Result of data_handle().}

\item{target}{Name of the target column.Default is "target"}
}
\value{
TaskClassif,a task base on mlr3
}
\description{
Create Task
}
\examples{
\donttest{
    data(test)
    data = data_handle(test)
    task = task_creat(data)}
}
